changed: put play count videodatabase methods in separate file
